requirements.txt
absl-py==2.3.1
attrs==20.3.0
beautifulsoup4==4.13.4
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
check50==3.3.11
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
Cython==3.2.0
dbus-python==1.2.18
deep-sort-realtime==1.3.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.19.1
flake8==7.3.0
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
gdown==5.2.0
grpcio==1.76.0
h5py==3.15.1
httplib2==0.20.2
idna==3.3
ImageIO==2.37.2
importlib-metadata==4.6.4
isort==4.3.21
jeepney==0.7.1
jellyfish==0.11.2
Jinja2==3.1.6
keyring==23.5.0
kiwisolver==1.4.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lib50==3.0.12
libvirt-python==8.0.0
Markdown==3.10
MarkupSafe==3.0.2
matplotlib==3.10.7
mccabe==0.7.0
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==2.2.6
oauthlib==3.2.0
opencv-python==4.12.0.88
packaging==25.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.1
polars-runtime-32==1.35.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.14.0
pyflakes==3.4.0
pygame==2.6.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.2.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.25.1
scipy==1.15.3
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.7
ssh-import-id==5.11
sympy==1.14.0
systemd-python==234
tb-nightly==2.21.0a20251023
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
tomli==2.3.0
torch==2.9.0+cpu
torchaudio==2.9.0+cpu
torchvision==0.24.0+cpu
tqdm==4.67.1
typing_extensions==4.14.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.226
ultralytics-thop==2.0.18
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
zipp==1.0.0
